+++ src/main/java/de/samply/teiler/utils/EnvironmentUtils.java
@@ -1,8 +1,9 @@
 package de.samply.teiler.utils;
 
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.core.env.AbstractEnvironment;
 import org.springframework.core.env.EnumerablePropertySource;
-import org.springframework.core.env.MutablePropertySources;
 
 import java.util.Arrays;
 import java.util.Map;
@@ -11,10 +12,16 @@ import java.util.function.Function;
 
 public class EnvironmentUtils {
 
+    private final static Logger logger = LoggerFactory.getLogger(EnvironmentUtils.class);
+
     public static void addKeyValuesFromEnvironment(AbstractEnvironment environment, Function<String, Boolean> isKeyToBeAdded, BiConsumer<String, String> addKeyValue) {
+        logger.info("Reading System Environment...");
         addKeyValuesFromEnvironment(environment.getSystemEnvironment(), isKeyToBeAdded, addKeyValue);
+        logger.info("Reading System Properties...");
         addKeyValuesFromEnvironment(environment.getSystemProperties(), isKeyToBeAdded, addKeyValue);
+        logger.info("Reading Property Sources...");
         addKeyValuesFromEnvironmentPropertySources(environment, isKeyToBeAdded, addKeyValue);
+        logger.info("Reading of environment variables finished.");
     }
 
     private static void addKeyValuesFromEnvironment(Map<String, Object> keyValues, Function<String, Boolean> isKeyToBeAdded, BiConsumer<String, String> addKeyValue) {